+++ hutool/extra/__init__.py
@@ -3,8 +3,10 @@ from .epub import EbooklibEpub, Epub, EpubFactory, MkEpub, PyPub3Epub
 from .pinyin import PinyinUtil
 from .qr_code import QrCodeUtil
 from .template import TemplateUtil
+from .tts import AsyncTtsUtil, TtsUtil, TtsVoice
 
 __all__ = [
+    "AsyncTtsUtil",
     "EbooklibEpub",
     "EmojiUtil",
     "Epub",
@@ -14,4 +16,6 @@ __all__ = [
     "PyPub3Epub",
     "QrCodeUtil",
     "TemplateUtil",
+    "TtsUtil",
+    "TtsVoice",
 ]